refactor(webapi): split chathub file to partial
